backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.3

# Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
urllib3==2.3.0

# Data Processing
pandas==2.2.3
numpy==2.1.3
tqdm==4.67.1

# Vietnamese NLP
underthesea==8.3.0
pyvi==0.1.1

# Deep Learning & Transformers
torch==2.9.1
transformers==4.57.3

# Text Analysis & Visualization
wordcloud==1.9.5
scikit-learn==1.6.1
matplotlib==3.10.0

# Sentiment Analysis
vaderSentiment==3.3.2

# db
chromadb==1.4.0
sentence-transformers==5.2.0
google-generativeai==0.8.6

# wsl --shutdown
# optimize-vhd -Path "C:\Users\<Tên_Bạn>\AppData\Local\Docker\wsl\data\ext4.vhdx" -Mode Full